Move update logic to `GitHubClient` process manager.
