--- mentions/src/main/proto/spine_examples/pingh/mentions/auto_update_mentions.proto
@@ -1,57 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-syntax = "proto3";
-
-package spine_examples.pingh.mentions;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.pingh.spine.io";
-option java_package = "io.spine.examples.pingh.mentions";
-option java_outer_classname = "AutoUpdateMentionsProto";
-option java_multiple_files = true;
-
-import "spine_examples/pingh/mentions/identifiers.proto";
-import "google/protobuf/timestamp.proto";
-
-// The process of automatically updating user mentions.
-//
-// This process periodically sends `UpdateMentionsFromGitHub` commands with a predefined delay.
-// The process starts when the `GitHubClient` obtains a `PersonalAccessToken`.
-//
-message AutoUpdateMentions {
-    option (entity) = { kind: PROCESS_MANAGER };
-
-    // The ID of the auto update mentions process.
-    GitHubClientId id = 1;
-
-    // The time when a mention update was last requested.
-    //
-    // If no requests have been made yet, this field is empty.
-    //
-    google.protobuf.Timestamp when_last_requested = 2;
-}
